requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
streamlit==1.49.0
chromadb==1.0.20
ollama==0.5.3
PyMuPDF==1.26.3
pytesseract==0.3.13
Pillow==11.3.0
python-multipart==0.0.20
requests==2.32.5
